pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "bev_multimae"
version = "0.0.1"
description = "A unified masked autoencoder architecture for fusing BEV radar representations and camera features projected into BEV space, inspired by Voxel-MAE, BEV-MAE, and MultiMAE."
authors = [
  { name = "Vebj\u00f8rn S. Skre", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "fastapi==0.115.6",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "huggingface-hub>=1.4.1",
    "datasets>=4.5.0",
    "matplotlib>=3.10.8",
    "mcap>=1.3.1",
    "invoke>=2.2.0",
    "mcap-ros2-support>=0.5.7",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "opencv-python>=4.13.0.92",
    "open3d>=0.19.0",
    "transformers>=5.3.0",
    "mmengine>=0.10.7",
    "depth-pro",
    "timm==1.0.3",
    "einops>=0.8.2",
    "wandb>=0.25.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
]


[tool.ruff]
line-length = 120

[tool.uv]
environments = ["sys_platform == 'linux'"]
override-dependencies = ["numpy>=1.24"]

[tool.uv.sources]
depth-pro = { git = "https://github.com/apple/ml-depth-pro.git" }

[tool.coverage.run]
omit = ["tests/*"]
